Move bundleParams to its own file.
