+++ comp/core/bundle.go
@@ -14,84 +14,11 @@ package core
 
 import (
 	"github.com/DataDog/datadog-agent/comp/core/config"
-	"github.com/DataDog/datadog-agent/comp/core/internal"
 	"github.com/DataDog/datadog-agent/comp/core/log"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
 // team: agent-shared-components
-
-// BundleParams defines the parameters for this bundle.
-type BundleParams = internal.BundleParams
-
-func CreateAgentBundleParams(confFilePath string, configLoadSecrets bool, options ...func(*BundleParams)) BundleParams {
-	params := CreateBundleParams(options...)
-	params.ConfFilePath = confFilePath
-	params.ConfigLoadSecrets = configLoadSecrets
-	return params
-}
-
-func CreateBundleParams(options ...func(*BundleParams)) BundleParams {
-	bundleParams := BundleParams{}
-	for _, o := range options {
-		o(&bundleParams)
-	}
-	return bundleParams
-}
-
-func WithConfigName(name string) func(*BundleParams) {
-	return func(b *BundleParams) {
-		b.ConfigName = name
-	}
-}
-
-func WithConfigMissingOK(v bool) func(*BundleParams) {
-	return func(b *BundleParams) {
-		b.ConfigMissingOK = v
-	}
-}
-
-func WithConfigLoadSysProbe(v bool) func(*BundleParams) {
-	return func(b *BundleParams) {
-		b.ConfigLoadSysProbe = v
-	}
-}
-
-func WithSecurityAgentConfigFilePaths(securityAgentConfigFilePaths []string) func(*BundleParams) {
-	return func(b *BundleParams) {
-		b.SecurityAgentConfigFilePaths = securityAgentConfigFilePaths
-	}
-}
-
-func WithConfigLoadSecurityAgent(configLoadSecurityAgent bool) func(*BundleParams) {
-	return func(b *BundleParams) {
-		b.ConfigLoadSecurityAgent = configLoadSecurityAgent
-	}
-}
-
-func WithConfFilePath(confFilePath string) func(*BundleParams) {
-	return func(b *BundleParams) {
-		b.ConfFilePath = confFilePath
-	}
-}
-
-func WithConfigLoadSecrets(configLoadSecrets bool) func(*BundleParams) {
-	return func(b *BundleParams) {
-		b.ConfigLoadSecrets = configLoadSecrets
-	}
-}
-
-func WithExcludeDefaultConfPath(excludeDefaultConfPath bool) func(*BundleParams) {
-	return func(b *BundleParams) {
-		b.ExcludeDefaultConfPath = excludeDefaultConfPath
-	}
-}
-
-func WithDefaultConfPath(defaultConfPath string) func(*BundleParams) {
-	return func(b *BundleParams) {
-		b.DefaultConfPath = defaultConfPath
-	}
-}
 
 // Bundle defines the fx options for this bundle.
 var Bundle = fxutil.Bundle(
